Extract ingest/crypto.js and aws.js{KWS} into @cumulus/common.
crypto.js is renamed key-pair-provider.js and moved from ingest package to
common pacakge.

KMS class is extracted from ingest/aws.js and moved to common, now accessed via
@cumulus/common/key-pair-provider.KMSProvider
