+++ packages/api/lambdas/bootstrap.js
@@ -19,7 +19,7 @@ const log = require('@cumulus/common/log');
 const pLimit = require('p-limit');
 const { dynamodb } = require('@cumulus/common/aws');
 const { inTestMode } = require('@cumulus/common/test-utils');
-const { DefaultProvider } = require('@cumulus/ingest/crypto');
+const { DefaultProvider } = require('@cumulus/common/key-pair-provider');
 const { User } = require('../models');
 const { Search, defaultIndexAlias } = require('../es/search');
 const mappings = require('../models/mappings.json');
