+++ packages/api/tests/serial/endpoints/granules.js
@@ -6,7 +6,7 @@ const sinon = require('sinon');
 const test = require('ava');
 const aws = require('@cumulus/common/aws');
 const { CMR } = require('@cumulus/cmrjs');
-const { DefaultProvider } = require('@cumulus/ingest/crypto');
+const { DefaultProvider } = require('@cumulus/common/key-pair-provider');
 const { randomString } = require('@cumulus/common/test-utils');
 const xml2js = require('xml2js');
 const { xmlParseOptions } = require('@cumulus/cmrjs/utils');
