+++ packages/cmrjs/package.json
@@ -36,6 +36,7 @@
     "lodash.get": "^4.4.2",
     "lodash.property": "^4.4.2",
     "public-ip": "^2.3.5",
+    "url-join": "^1.0.0",
     "xml2js": "^0.4.19"
   },
   "devDependencies": {
