+++ packages/common/index.js
@@ -11,3 +11,4 @@ exports.stepFunctions = require('./step-functions');
 exports.stringUtils = require('./string');
 exports.testUtils = require('./test-utils');
 exports.util = require('./util');
+exports.keyPairProvider = require('./key-pair-provider');
