+++ packages/common/package.json
@@ -62,6 +62,7 @@
     "lodash.range": "^3.2.0",
     "mocha": "^5.0.4",
     "mocha-junit-reporter": "^1.11.1",
+    "node-forge": "^0.7.1",
     "p-limit": "^2.0.0",
     "p-map": "^1.2.0",
     "p-retry": "^2.0.0",
